fix tests
Former-commit-id: 30e21b5e2d420513fe816d561fb709661e7f513e [formerly 0d29136bfcd8189a1f6e8c8f7690710d07e4bf19 [formerly c151c626cf14448a893d90dbf3f9fc0e70957159]]
Former-commit-id: 0d29136bfcd8189a1f6e8c8f7690710d07e4bf19
Former-commit-id: 5f7559b0f2df3cc74afa8acc508de9cd1f1e3b32

+++ gdsfactory/simulation/gmeep/write_sparameters_meep_mpi.py
@@ -121,7 +121,7 @@ def write_sparameters_meep_mpi(
         logger.info(f"Simulation {filepath!r} already exists")
         return filepath
 
-    if overwrite:
+    if filepath.exists() and overwrite:
         filepath.unlink()
 
     # Save the component object to simulation for later retrieval
